Move User Model To Appropriate Directory
