+++ Configuration/Extbase/Persistence/Classes.php
@@ -3,10 +3,10 @@
 return [
     \In2code\Femanager\Domain\Model\User::class => [
         'subclasses' => [
-            \BalatD\FriendlyCaptcha\Domain\Model\Domain\User::class
-        ]
+            '\BalatD\FriendlyCaptcha\Domain\Model\User' => \BalatD\FriendlyCaptcha\Domain\Model\User::class,
+        ],
     ],
-    \BalatD\FriendlyCaptcha\Domain\Model\Domain\User::class => [
+    \BalatD\FriendlyCaptcha\Domain\Model\User::class => [
         'tableName' => 'fe_users',
         'recordType' => 0,
     ],
